add trajectory.py, move things around in snapshots.py
